Implement estimates of covariance of SSM-parameters, and spectral estimator
The covariance of the parameter estimates of the SSM A is implemented.
This is mostly a copy of Lindas code.

In addition, this commit implements an estimator of the power spectrum
based on VAR parameters. In contrast to the equation given in the paper
this implementation is correct for the multivariate case.
On branch emfitCovAndSpec
	modified:   fit/emfit.cpp
no changes added to commit (use "git add" and/or "git commit -a")