Return proper (hopefully) estimate errors.
Most probably, the diagonal of the big V matrix in
emfitEstimateErrors.cpp hold the uncertainties of the
estimated parameteres. Most probably, the orderin of
the diagonal entries are A11(1), A12(1), ..., Add(t)
followed by the errors for Q and R. I right now
double-check this with Linda. Strangely, the values are
sometimes negative, but this might be due to the bad fit
conditions I used for testing (only one em iteration).
The errors do not come from a quadaric form, but from the
expected vaule of the second derivitive of the likelihood,
the inverse thereof. So, I can imagin, that bad things
might happen away from all sensible parameters.